agrego validaciones de sesion y archivo requirements.txt #4

+++ backend/db/insert_datos.sql
@@ -1,4 +1,4 @@
-DELETE FROM usuarios;
+DELETE FROM usuarios WHERE email != 'admin@puertohermoso.com';
 
 DELETE FROM mesa;
